Fixing set filter bug

+++ R2U2_C/src/AT/signal_set_filters/filter_exactly_one_of.c
@@ -1,10 +1,11 @@
 
 #include "filter_exactly_one_of.h"
 
-bool filter_exactly_one_of(bool *set, size_t len)
+bool filter_exactly_one_of(bool *set, uint8_t len)
 {
     bool res = 0;
-    int i;
+    uint8_t i;
+
     for(i = 0; i < len; i++) {
         if(res && set[i]) return 0;
         res = res || set[i];
