+++ R2U2_C/src/AT/signal_set_filters/filter_exactly_one_of.h
@@ -7,6 +7,6 @@
 
 #include "R2U2Config.h"
 
-bool filter_exactly_one_of(bool *set, size_t len);
+bool filter_exactly_one_of(bool *set, uint8_t len);
 
 #endif
